+++ packages/react/src/components/Dropdown/Dropdown.tsx
@@ -185,7 +185,7 @@ export interface DropdownProps
   renderItem?: ShorthandRenderFunction
 
   /**
-   * A custom render function for the selected item.
+   * A custom render function for the selected item. Only applicable with the multiple prop.
    *
    * @param {React.ReactType} Component - The computed component for this slot.
    * @param {object} props - The computed props for this slot.