+++ packages/react/test/specs/components/Dropdown/Dropdown-test.tsx
@@ -8,7 +8,6 @@ import DropdownSelectedItem from 'src/components/Dropdown/DropdownSelectedItem'
 import { isConformant } from 'test/specs/commonTests'
 import { findIntrinsicElement, mountWithProvider } from 'test/utils'
 import { ReactWrapper, CommonWrapper } from 'enzyme'
-import Button from 'src/components/Button/Button'
 
 jest.dontMock('keyboard-key')
 jest.useFakeTimers()
@@ -1444,38 +1443,6 @@ describe('Dropdown', () => {
   })
 
   describe('renderSelectedItem', () => {
-    it('calls renderSelectedItem', () => {
-      const renderSelectedItem = jest.fn((button, props) => (
-        <Button {...props} content={`${props.content} (selected)`} />
-      ))
-      const wrapper = mountWithProvider(
-        <Dropdown items={items} renderSelectedItem={renderSelectedItem} />,
-      )
-      const triggerButton = getTriggerButtonWrapper(wrapper)
-
-      triggerButton.simulate('click')
-      const firstItem = getItemAtIndexWrapper(wrapper)
-      firstItem.simulate('click', { nativeEvent: { stopImmediatePropagation: jest.fn() } })
-
-      expect(renderSelectedItem).toBeCalled()
-    })
-
-    it('renders custom text', () => {
-      const renderSelectedItem = jest.fn((button, props) => (
-        <Button {...props} content={`${props.content} (selected)`} />
-      ))
-      const wrapper = mountWithProvider(
-        <Dropdown items={items} renderSelectedItem={renderSelectedItem} />,
-      )
-      const triggerButton = getTriggerButtonWrapper(wrapper)
-
-      triggerButton.simulate('click')
-      const firstItem = getItemAtIndexWrapper(wrapper)
-      firstItem.simulate('click', { nativeEvent: { stopImmediatePropagation: jest.fn() } })
-
-      expect(triggerButton.getDOMNode().textContent).toBe('item1 (selected)')
-    })
-
     it('calls renderSelectedItem in multiple selection', () => {
       const renderSelectedItem = jest.fn((SelectedItem, props) => (
         <SelectedItem {...props} header={`${props.header} (selected)`} />